+++ model/model_action_reward_rule.go
@@ -12,7 +12,7 @@ type ActionRewardRule struct {
 	ID              uint      `gorm:"primary_key" json:"id"`
 	ActionName      string    `json:"action_name"`
 	DisplayName     string    `json:"display_name"`
-	Amount          uint      `json:"amount"`
+	Amount          float64   `json:"amount"`
 	Limit           uint      `json:"limit"`
 	UpdatedAt       time.Time `json:"updated_at"`
 	CreatedAt       time.Time `json:"created_at"`
